Replace Microsoft.Owin.Security.Cookies.Interop by Microsoft.Owin.Security.Interop
