--- src/Microsoft.Owin.Security.Cookies.Interop/project.json
@@ -1,18 +1,0 @@
-{
-    "version": "1.0.0-*",
-    "compilationOptions": {
-        "warningsAsErrors": true,
-        "keyFile": "../../tools/Key.snk"
-    },
-    "description": "A compatibility layer for sharing authentication tickets between Microsoft.Owin.Security.Cookies and Microsoft.AspNet.Authentication.Cookies.",
-    "dependencies": {
-    },
-    "frameworks": {
-        "net451": {
-            "dependencies": {
-                "Microsoft.AspNet.DataProtection.Extensions": "1.0.0-*",
-                "Microsoft.Owin.Security.Cookies": "3.0.1"
-            }
-        }
-    }
-}
