+++ src/Microsoft.Owin.Security.Interop/project.json
@@ -1,0 +1,15 @@
+﻿{
+    "version": "1.0.0-*",
+    "compilationOptions": {
+        "warningsAsErrors": true,
+        "keyFile": "../../tools/Key.snk"
+    },
+    "description": "A compatibility layer for sharing authentication tickets between Microsoft.Owin.Security and Microsoft.AspNet.Authentication.",
+    "dependencies": {
+        "Microsoft.AspNet.DataProtection.Extensions": "1.0.0-*",
+        "Microsoft.Owin.Security": "3.0.1"
+    },
+    "frameworks": {
+        "net451": { }
+    }
+}
